requirements.txt
absl-py==2.1.0
appdirs==1.4.4
astroid==3.1.0
astunparse==1.6.3
biopython==1.83
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
docker==7.0.0
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
isort==5.13.2
joblib==1.4.0
keras==3.2.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mymetal==1.0.9
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
progress==1.6
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybiolib==1.1.2004
Pygments==2.17.2
pylint==3.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sklearn==0.0.post12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
tomlkit==0.12.4
typed-ast==1.5.5
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
